chore(client): clean code
